moved blockservice out
License: MIT
